+++ common/src/main/java/net/sensible_survival/compat/CompatFeatures.java
@@ -1,0 +1,7 @@
+package net.sensible_survival.compat;
+
+public class CompatFeatures {
+    public static void initialize() {
+        FTBTeamsCompat.init();
+    }
+}