+++ common/src/main/java/net/sensible_survival/mixin/LivingEntityMixin.java
@@ -1,6 +1,6 @@
-package net.durability_tweaks.mixin;
+package net.sensible_survival.mixin;
 
-import net.durability_tweaks.DurabilityTweaksMod;
+import net.sensible_survival.SensibleSurvivalMod;
 import net.minecraft.entity.LivingEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
@@ -15,7 +15,7 @@ public class LivingEntityMixin {
     private static final Random DURABILITY_RNG = new Random();
     @ModifyVariable(method = "damageEquipment", at = @At("HEAD"), ordinal = 0, argsOnly = true)
     private float modifyExhaustion(float amount) {
-        var config = DurabilityTweaksMod.getConfig();
+        var config = SensibleSurvivalMod.getConfig();
         if (config.equipment_takes_damage_chance < 1F
                 && DURABILITY_RNG.nextFloat() > config.equipment_takes_damage_chance) {
             return 0F;